Bump version 0.28.0.post1 → 0.28.1
Bug fixes
---------

* Fix Fief CLI not correctly configured

0.28.0 notes
------------

Multi-workspaces removal
------------------------

As we announced last month in our [blog post](https://polar.sh/fief-dev/posts/the-future-of-fief), we are removing the multi-workspaces feature to focus on a better and easier self-hosting experience.

As such, if you are currently self-hosting a Fief server, you'll need to perform a few migrations steps, as detailed here: https://docs.fief.dev/migration

Basically, it only consists of moving the data from a sub-schema to the main one.

Fief Cloud sunset
-----------------

Fief Cloud will also be sunsetted. Starting today, it won't be possible to create new workspaces.

For existing workspaces, we will keep the service running until at least **1st of July 2024**, but we encourage you to migrate to a self-hosted solution as soon as possible.

To ease the transition, we've created a dedicated migration tool and guide: https://docs.fief.dev/migration

Bug fixes and improvements
--------------------------

* An admin role and permission is now automatically created at instance startup
    * This role is required to access the admin dashboard
* Fix #228: default roles are now granted synchronously upon registration
* Upgrade official image to Python 3.12